+++ tests/unit/validator/rules/test/MimetypeTest.php
@@ -1,0 +1,65 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\tests\unit\validator\rules\file;
+
+use mako\tests\TestCase;
+use mako\validator\rules\file\Mimetype;
+use SplFileInfo;
+
+/**
+ * @group unit
+ */
+class AlphaTest extends TestCase
+{
+	/**
+	 *
+	 */
+	public function setUp()
+	{
+		if(function_exists('finfo_open') === false)
+		{
+			$this->markTestSkipped("The fileinfo extension hasn't been enabled.");
+		}
+	}
+
+	/**
+	 *
+	 */
+	public function testValidatesWhenEmpty()
+	{
+		$rule = new Mimetype;
+
+		$this->assertFalse($rule->validateWhenEmpty());
+	}
+
+	/**
+	 *
+	 */
+	public function testWithValidValue()
+	{
+		$rule = new Mimetype;
+
+		$rule->setParameters(['image/png']);
+
+		$this->assertTrue($rule->validate(new SplFileInfo(__DIR__ . '/fixtures/png.png'), []));
+	}
+
+	/**
+	 *
+	 */
+	public function testWithInvalidValue()
+	{
+		$rule = new Mimetype;
+
+		$rule->setParameters([['text/plain', 'application/json']]);
+
+		$this->assertFalse($rule->validate(new SplFileInfo(__DIR__ . '/fixtures/png.png'), []));
+
+		$this->assertSame('The foobar must be a file of type: text/plain, application/json.', $rule->getErrorMessage('foobar'));
+	}
+}